decoupled public and private user info
