+++ server/schema/types/root_query_type.js
@@ -2,8 +2,7 @@ const graphql = require('graphql');
 
 const { GraphQLObjectType, GraphQLString } = graphql;
 const UserType = require('./user_type');
-const UserTestType = require('./user_test_type');
-const { verifyToken, readTest, getUser } = require('../../services/auth');
+const { getUser } = require('../../services/auth');
 
 const RootQueryType = new GraphQLObjectType({
   name: 'RootQueryType',
@@ -21,23 +20,6 @@ const RootQueryType = new GraphQLObjectType({
           .catch((e) => {
             console.error(e);
             rej(e);
-          });
-        });
-      }
-    },
-    userId: {
-      type: UserTestType,
-      args: { token: { type: GraphQLString } },
-      resolve: (parentValue, { token }) => {
-        return new Promise((res, rej) => {
-          verifyToken(token)
-          .then((uid) => {
-            readTest(uid);
-            res({ id: uid });
-          })
-          .catch((e) => {
-            console.error(e);
-            rej();
           });
         });
       }
