--- server/schema/types/user_test_type.js
@@ -1,15 +1,0 @@
-const graphql = require('graphql');
-
-const {
-  GraphQLObjectType,
-  GraphQLID,
-} = graphql;
-
-const UserTestType = new GraphQLObjectType({
-  name: 'UserTestType',
-  fields: () => ({
-    id: { type: GraphQLID },
-  })
-});
-
-module.exports = UserTestType;
