+++ server/schema/types/user_type.js
@@ -14,14 +14,8 @@ const UserType = new GraphQLObjectType({
     email: { type: GraphQLString },
     username: { type: GraphQLString },
     photo: { type: GraphQLString },
-    location: { type: new GraphQLList(LocationType) },
-    friends: { type: new GraphQLList(UserType) },
-    groups: { type: new GraphQLList(GroupType) }
+    friends: { type: new GraphQLList(GraphQLID) },
   })
 });
 
 module.exports = UserType;
-
-const LocationType = require('./location_type');
-const GroupType = require('./group_type');
-
